requirements.txt
streamlit>=1.30.0
lancedb>=0.4.0
langchain-google-genai>=1.0.0
langchain>=0.2.0
python-dotenv>=1.0.0
